+++ hamcrest-matchers/guava-hamcrest-matchers/src/main/java/it/ozimov/cirneco/hamcrest/guava/collect/IsMultimapWithKeySetSize.java
@@ -18,7 +18,7 @@ public class IsMultimapWithKeySetSize<K> extends TypeSafeMatcher<Multimap<K, ?>>
     }
 
     /**
-     * Creates a matcher for {@linkplain Multimap} matching when the examined object has an empty key set.
+     * Creates a matcher for {@linkplain Multimap} matching when the examined object has an withoutRepetitions key set.
      */
     public static <K> Matcher<Multimap<K, ?>> emptyKeySet() {
         return new IsMultimapWithKeySetSize(0);