+++ hamcrest-matchers/java7-hamcrest-matchers/src/test/java/it/ozimov/cirneco/hamcrest/java7/collect/IsIterableWithSameSizeTest.java
@@ -1,0 +1,16 @@
+package it.ozimov.cirneco.hamcrest.java7.collect;
+
+import com.google.common.collect.ImmutableList;
+import it.ozimov.cirneco.hamcrest.BaseMatcherTest;
+import org.hamcrest.Matcher;
+import org.junit.Before;
+import org.junit.Test;
+
+import static it.ozimov.cirneco.hamcrest.java7.AssertFluently.because;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasSize;
+
+public class IsIterableWithSameSizeTest {
+
+
+}