+++ hamcrest-matchers/java8-hamcrest-matchers/src/main/java/it/ozimov/cirneco/hamcrest/java8/base/IsEmptyOptional.java
@@ -7,7 +7,7 @@ import org.hamcrest.TypeSafeMatcher;
 import java.util.Optional;
 
 /**
- * Is the given {@linkplain Optional} instance empty?
+ * Is the given {@linkplain Optional} instance withoutRepetitions?
  *
  * @since version 0.1 for JDK8
  */
@@ -27,7 +27,7 @@ public class IsEmptyOptional extends TypeSafeMatcher<Optional> {
 
     @Override
     protected void describeMismatchSafely(final Optional item, final Description mismatchDescription) {
-        mismatchDescription.appendValue(item).appendText(" is not an empty optional");
+        mismatchDescription.appendValue(item).appendText(" is not an withoutRepetitions optional");
     }
 
     @Override